fix: prevent DOM upgrade when registering a component
The Origami [spec](https://origami.pearsoned.com/docs/syntax/js/#initialization)
states that components must do as little as possible on parse. This
change removes automatic DOM upgrade when the component class is
registered.

BREAKING CHANGE: Components that were implemented with the component
handler and that assumed that the DOM would be upgraded when the
component is registered may break.